@@ -13,8 +13,8 @@
     "carbon-icons": "^7.0.7",
     "civic-info": "0.0.5",
     "material-table": "^1.69.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.1",
