@@ -4,8 +4,7 @@ import RepForm from "./RepForm";
 import Welcome from "./Welcome";
 import Navbar from "./Navbar";
 import VotingRecords from "./VotingRecords";
-import { createMuiTheme } from "@material-ui/core/styles";
-import { ThemeProvider } from "@material-ui/styles";
+import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
 
 const theme = createMuiTheme({
   typography: {
